Update package version and dependencies
Updated the package version from 0.2.15 to 0.3.0 in version.py and setup.py. This was necessary to reflect the latest changes in the package. The dependencies numpy, scipy, and pandas were also updated to the versions 1.26.0, 1.11.3, and 2.1.3 respectively. These updates are needed to ensure compatibility with the latest versions of these libraries.